feat: Add player count to users component (#111)
looks like this:

![image](https://github.com/MaSch0212/minigolf-friday/assets/9435005/aab44313-89f7-46da-8bd0-26c50226754a)

it now will display the admin users on the top of the page and the
"normal" users below them. each "section" is alphabetically ordered.
also add the functionality to insert the search value directly to the
name entry of the "create new User" view.
With this, you can search for a user and if no result is found, just
click add user and the name will already be filled in.

Closes: #106